survey user mgmt

+++ api-server/src/http/controllers/admin/survey-mgmt.controller.ts
@@ -1,11 +1,12 @@
 import { Request, Response, NextFunction } from 'express';
 import { Op } from 'sequelize';
+import { UserResponse } from '@common/types/api/admin/users';
 import Survey from '@/db/models/system/survey';
 import User from '@/db/models/system/user';
 import UserRole from '@/db/models/system/user-role';
 import NotFoundError from '@/http/errors/not-found.error';
 import userResponse from '@/http/responses/admin/user.response';
-import { surveyStaff, surveySupport } from '@/services/acl.service';
+import { surveyMgmt } from '@/services/acl.service';
 
 export default {
   async list(req: Request, res: Response, next: NextFunction): Promise<void> {
@@ -17,21 +18,17 @@ export default {
       return;
     }
 
-    const staff = await User.paginate({
+    const users = await User.paginate<UserResponse>({
       req,
       columns: ['name', 'email', 'simpleName'],
-      include: [
-        { model: UserRole, where: { role: [surveyStaff(surveyId), surveySupport(surveyId)] } },
-      ],
+      include: [{ model: UserRole, where: { role: surveyMgmt(surveyId) } }],
       transform: userResponse,
     });
 
-    const users = await User.findAll({ where: { email: { [Op.ne]: null } } });
-
-    res.json({ ...staff, refs: { users } });
+    res.json(users);
   },
 
-  async store(req: Request, res: Response, next: NextFunction): Promise<void> {
+  async available(req: Request, res: Response, next: NextFunction): Promise<void> {
     const { surveyId } = req.params;
     const survey = await Survey.findByPk(surveyId);
 
@@ -40,18 +37,29 @@ export default {
       return;
     }
 
-    const { userId } = req.body;
-
-    await UserRole.destroy({
-      where: { userId, role: [surveyStaff(surveyId), surveySupport(surveyId)] },
+    const users = await User.findAll({
+      where: { email: { [Op.ne]: null } },
+      include: [{ model: UserRole, where: { role: { [Op.notIn]: surveyMgmt(surveyId) } } }],
     });
 
-    const roles = req.body.roles as string[];
-    if (roles.length) {
-      const roleRecords = roles.map((role) => ({ userId, role }));
-      await UserRole.bulkCreate(roleRecords);
+    res.json({ data: users.map(userResponse) });
+  },
+
+  async update(req: Request, res: Response, next: NextFunction): Promise<void> {
+    const { surveyId, userId } = req.params;
+    const survey = await Survey.findByPk(surveyId);
+    const user = await User.findOne({ where: { email: { [Op.ne]: null } } });
+
+    if (!survey || !user) {
+      next(new NotFoundError());
+      return;
     }
 
-    res.status(201).json();
+    await UserRole.destroy({ where: { userId, role: surveyMgmt(surveyId) } });
+
+    const roles = (req.body.roles as string[]).map((role) => ({ userId, role }));
+    if (roles.length) await UserRole.bulkCreate(roles);
+
+    res.json();
   },
 };
